homework2/requirements.txt
asgiref==3.11.1
brotli==1.2.0
click==8.3.1
dj-database-url==3.1.2
Django==6.0.2
djangorestframework==3.16.1
gunicorn==25.1.0
h11==0.16.0
iniconfig==2.3.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
pytz==2026.1.post1
sqlparse==0.5.5
uvicorn==0.41.0
whitenoise==6.12.0
